2.ComputerVision/1.cv2/2.cv_project_streamlit/requirements.txt
streamlit
opencv-python
Pillow
numpy
